refactor into BoltzService class
